backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.0
hashlib-compat==1.0.0

# Additional dependencies
pydantic[email]==2.5.3
